refactor(charts): rename SQL files to include Artists for consistency
