+++ app/src/components/Charts/LabCharts/Top10Race/Top10Artists.sql
