+++ app/src/components/Charts/LabCharts/Top10Race/query.ts
@@ -1,5 +1,5 @@
 import { TABLE } from '../../../../db/queries/constants'
-import sqlTop10Race from './Top10Race.sql?raw'
+import sqlTop10Race from './Top10Artists.sql?raw'
 import sqlTop10Tracks from './Top10Tracks.sql?raw'
 import sqlTop10Albums from './Top10Albums.sql?raw'
